Add package rbac for RBAC observer client
Package rbac provides a k8s client that can observe and record the RBAC
permissions for the API calls and provide Role and Cluster Role
permissions required to run the code.
NOTE: Using the controller-runtime fake client with the rbac client
fails because the fake client doesn't implement a RESTMapper. RESTMapper
is required by the RBACClient to figure out the scope of the resource
objects, namespaced or cluster resource. It works properly in envtest
setup.