sync: Run the sync func at initialization
Run the sync func at initialization, before setting up the sync
ticker. This will ensure the sync is run at controller startup.
Also adds an option to set an initial sync delay. By default the delay
is set to 10 seconds to allow the controller components to initialize
before running the sync. This value can be overridden but not set to
zero. Setting it to zero adds the risk of running the sync before the
controller components are ready, resulting in cache related controller
errors.