declarative: DeepCopy FileSystem in NewBuilder
Adds DeepCopy functions for deep copying FileSystems. This is needed to
be able to create a copy of the filesystem before modifying the files in
the builder.
The NewBuilder function performs a deep copy of the given filesystem and
works on a copy of the filesystem by default.
This helps protect against various filesystem related conflicts, race
issues, conflicting persistent changes, non-idempotent transforms, etc.