Add support for informers with external objects
Add cache and source packages that are modifications of the
controller-runtime cache and source packages to support client-go
informers based cache that can collect objects from non-k8s API servers.

This enables writing controllers that can reconcile non-k8s objects
backed by the same informer based cache that k8s controllers use. The
cache can also be shared across multiple controllers, forming shared
informers. All the events in the cache are shared with all the
controllers that are interested in a particular object.
Unlike the cache created by the controller-manager, this cache needs to
be created separately when the operator starts and explicitly started to
start all the informers that belong to the cache. Because the cache
interface embeds Informer interface which has a Start(ctx) method, the
cache can be added to the controller-manager to manager. This makes
clean shutdown of the informers along with the other informers of the
managed cache.

This also includes a ListWatcher type to help create a list watcher
function that's needed to create an informer.