Add example controller for shared informer cache
It adds 3 controllers with a shared cache. The cache is created with a
mock ListWatcher that populates the event channels with test data. For
real API servers, the ListWatcher can be replaced with an actual client
implementation. This shared cache is passed to all the controllers that
are interested in mocked API objects. When the cache informers are
started by the controller manager. Uses the same context, which ensure
that the informers are stopped when the controller-manager's context is
cancelled.

Controllers spaceinformer1 and spaceinformer2 watch the same object.
podinformer1 watches a different object. The controllers that share
the same cache and watch the same object reconcile on the same events.
The CRDs used for these controllers need not be registered with the k8s
API server. To achieve this, a different kind event source is used along
with the external cache.