Add singleton package and webhook function
Introduces singleton package to help implement of any object kind. This
can be used in controllers directly or in admission controllers. The
webhook function ValidateSingletonCreate() can be used in any admission
controller to enable singleton of an object kind.

Also adds e2e tests for the webhook validate singleton function.