Composite opreator: set retry period on not ready
Operand's ensure call now returns a not ready error when the ready
check fails.

When operand ensure results in not ready error, the component isn't
ready. Set a retry period in the composite operator ensure result to
provide some time for the component to be ready, instead of retrying
immediately.