declarative/kubectl: Allow setting IOStreams
Update the declarative kubectl package to allow setting IOStreams in the
applier and deleter. By default the applier and deleter write to stdout
and stderr, in addition to returning error. This change allows
configuring the writers by passing IOStreams.

The applier implementation in the upstream
kubebuilder-declarative-pattern doesn't allow configuring the IOStreams.
This change copies the applier implementation and adds the necessary
options to allow configuring the IOStreams.